requirements.txt
PyQt6>=6.5.0
telethon>=1.35.0
cryptg>=0.4.0
qasync>=0.28.0
aiosqlite>=0.19.0
faker>=20.0.0
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0
